requirements.txt
openai
groq
python-dotenv
pydantic
rich
streamlit